move isvmRunning and isvmShutdown from lib/qubinode_ocp4.sh to lib/qubinode_utils.sh

+++ lib/qubinode_utils.sh
@@ -13,24 +13,27 @@ function qubinode_project_cleanup () {
         FILES=("${FILES[@]}" "$vault_vars_file" "$vars_file")
     fi
 
-    # Delete OpenShift files
-    openshift_product=$(awk '/^product:/ {print $2}' "${project_dir}/playbooks/vars/ocp3.yml")
-    if [[ ${openshift_product} == "ocp3" ]]; then
-      FILES=("${FILES[@]}" "$ocp3_vars_files")
-    elif [[ ${openshift_product} == "okd3" ]]; then
-      FILES=("${FILES[@]}" "$okd3_vars_files")
-    fi
+    # Delete OpenShift 3 files
+    if [ -f ${project_dir}/playbooks/vars/ocp3.yml ]
+    then 
+        openshift_product=$(awk '/^product:/ {print $2}' "${project_dir}/playbooks/vars/ocp3.yml")
+        if [[ ${openshift_product} == "ocp3" ]]; then
+          FILES=("${FILES[@]}" "$ocp3_vars_files")
+        elif [[ ${openshift_product} == "okd3" ]]; then
+          FILES=("${FILES[@]}" "$okd3_vars_files")
+        fi
 
-    if [ ${#FILES[@]} -eq 0 ]
-    then
-        echo "Project directory: ${project_dir} state is already clean"
-    else
-        for f in $(echo "${FILES[@]}")
-        do
-            test -f $f && rm $f
-            echo "purged $f"
+        if [ ${#FILES[@]} -eq 0 ]
+        then
+            echo "Project directory: ${project_dir} state is already clean"
+        else
+            for f in $(echo "${FILES[@]}")
+            do
+                test -f $f && rm $f
+                echo "purged $f"
 
-        done
+            done
+        fi
     fi
 
    echo "Removing playbook vars"
@@ -98,5 +101,13 @@ convertB_human() {
 
     DISK_SIZE_HUMAN=$(printf "%d %s\n" $NUMBER $DESIG)
     DISK_SIZE_COMPARE=$(printf "%d %s\n" $NUMBER)
+}
+
+function isvmRunning () {
+    sudo virsh list |grep $vm|awk '/running/ {print $2}'
+}
+
+function isvmShutdown () {
+    sudo virsh list --all | grep $vm| awk '/shut/ {print $2}'
 }
